Fix empty download bug if URL contains page number
If the URL provided by the user ends with the page number of the
Issuu document (e.g. https://issuu.com/user/docs/doc/18) then
the script would complete successfully but no image would be
downloaded and the pdf would be empty as a result.

This commit fixes the issue so that...
https://issuu.com/user/docs/doc/18/
is converted to...
https://issuu.com/user/docs/doc
before proceeding with the download.